QueryMonthlyBillResponse add roundDownDiscount.

+++ services/bssopenapi/create_resource_package.go
@@ -89,7 +89,7 @@ type CreateResourcePackageRequest struct {
 type CreateResourcePackageResponse struct {
 	*responses.BaseResponse
 	RequestId string                      `json:"RequestId" xml:"RequestId"`
-	OrderId   int                         `json:"OrderId" xml:"OrderId"`
+	OrderId   int64                       `json:"OrderId" xml:"OrderId"`
 	Success   bool                        `json:"Success" xml:"Success"`
 	Code      string                      `json:"Code" xml:"Code"`
 	Message   string                      `json:"Message" xml:"Message"`
