+++ services/bssopenapi/struct_evaluate.go
@@ -17,22 +17,22 @@ package bssopenapi
 
 // Evaluate is a nested struct in bssopenapi response
 type Evaluate struct {
-	Id                 int    `json:"Id" xml:"Id"`
+	Id                 int64  `json:"Id" xml:"Id"`
 	GmtCreate          string `json:"GmtCreate" xml:"GmtCreate"`
 	GmtModified        string `json:"GmtModified" xml:"GmtModified"`
-	UserId             int    `json:"UserId" xml:"UserId"`
+	UserId             int64  `json:"UserId" xml:"UserId"`
 	UserNick           string `json:"UserNick" xml:"UserNick"`
 	OutBizId           string `json:"OutBizId" xml:"OutBizId"`
-	BillId             int    `json:"BillId" xml:"BillId"`
-	ItemId             int    `json:"ItemId" xml:"ItemId"`
+	BillId             int64  `json:"BillId" xml:"BillId"`
+	ItemId             int64  `json:"ItemId" xml:"ItemId"`
 	BillCycle          string `json:"BillCycle" xml:"BillCycle"`
 	BizType            string `json:"BizType" xml:"BizType"`
-	OriginalAmount     int    `json:"OriginalAmount" xml:"OriginalAmount"`
-	PresentAmount      int    `json:"PresentAmount" xml:"PresentAmount"`
-	CanInvoiceAmount   int    `json:"CanInvoiceAmount" xml:"CanInvoiceAmount"`
-	InvoicedAmount     int    `json:"InvoicedAmount" xml:"InvoicedAmount"`
-	OffsetCostAmount   int    `json:"OffsetCostAmount" xml:"OffsetCostAmount"`
-	OffsetAcceptAmount int    `json:"OffsetAcceptAmount" xml:"OffsetAcceptAmount"`
+	OriginalAmount     int64  `json:"OriginalAmount" xml:"OriginalAmount"`
+	PresentAmount      int64  `json:"PresentAmount" xml:"PresentAmount"`
+	CanInvoiceAmount   int64  `json:"CanInvoiceAmount" xml:"CanInvoiceAmount"`
+	InvoicedAmount     int64  `json:"InvoicedAmount" xml:"InvoicedAmount"`
+	OffsetCostAmount   int64  `json:"OffsetCostAmount" xml:"OffsetCostAmount"`
+	OffsetAcceptAmount int64  `json:"OffsetAcceptAmount" xml:"OffsetAcceptAmount"`
 	Status             int    `json:"Status" xml:"Status"`
 	OpId               string `json:"OpId" xml:"OpId"`
 	Name               string `json:"Name" xml:"Name"`
